helpers 2 on existing overlays
